Break apart the shadow demon ushering in a thousand years of darkness upon the world into separate pieces never to be combined, attempt 0
